+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+
+- filtered classroom list
+
 ## [4.0.0-beta.11] - 2022-11-08
 
 ### Added